Remove unused caches and move config into separate file
